feat(coupon-code-input): rename from `cart-discount-input`
